Move NonReturningFunctionInspectionTests for QuickFixes into relevant files under RD.QuickFixes
