+++ RubberduckTests/Inspections/NonReturningFunctionInspectionTests.cs
@@ -1,7 +1,6 @@
 using System.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Rubberduck.Inspections.Concrete;
-using Rubberduck.Inspections.QuickFixes;
 using Rubberduck.Parsing.Inspections.Resources;
 using Rubberduck.VBEditor.SafeComWrappers;
 using Rubberduck.VBEditor.SafeComWrappers.Abstract;
@@ -175,214 +174,6 @@ End Function";
             var inspectionResults = inspection.GetInspectionResults();
 
             Assert.AreEqual(1, inspectionResults.Count());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void NonReturningFunction_QuickFixWorks_Function()
-        {
-            const string inputCode =
-@"Function Foo() As Boolean
-End Function";
-
-            const string expectedCode =
-@"Sub Foo()
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new NonReturningFunctionInspection(state);
-            var inspectionResults = inspection.GetInspectionResults();
-
-            new ConvertToProcedureQuickFix(state).Fix(inspectionResults.First());
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void GivenFunctionNameWithTypeHint_SubNameHasNoTypeHint()
-        {
-            const string inputCode =
-@"Function Foo$()
-End Function";
-
-            const string expectedCode =
-@"Sub Foo()
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new NonReturningFunctionInspection(state);
-            var inspectionResults = inspection.GetInspectionResults();
-
-            new ConvertToProcedureQuickFix(state).Fix(inspectionResults.First());
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void NonReturningFunction_QuickFixWorks_FunctionReturnsImplicitVariant()
-        {
-            const string inputCode =
-@"Function Foo()
-End Function";
-
-            const string expectedCode =
-@"Sub Foo()
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new NonReturningFunctionInspection(state);
-            var inspectionResults = inspection.GetInspectionResults();
-
-            new ConvertToProcedureQuickFix(state).Fix(inspectionResults.First());
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void NonReturningFunction_QuickFixWorks_FunctionHasVariable()
-        {
-            const string inputCode =
-@"Function Foo(ByVal b As Boolean) As String
-End Function";
-
-            const string expectedCode =
-@"Sub Foo(ByVal b As Boolean)
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new NonReturningFunctionInspection(state);
-            var inspectionResults = inspection.GetInspectionResults();
-
-            new ConvertToProcedureQuickFix(state).Fix(inspectionResults.First());
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void GivenNonReturningPropertyGetter_QuickFixConvertsToSub()
-        {
-            const string inputCode =
-@"Property Get Foo() As Boolean
-End Property";
-
-            const string expectedCode =
-@"Sub Foo()
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new NonReturningFunctionInspection(state);
-            var inspectionResults = inspection.GetInspectionResults();
-
-            new ConvertToProcedureQuickFix(state).Fix(inspectionResults.First());
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void GivenNonReturningPropertyGetWithTypeHint_QuickFixDropsTypeHint()
-        {
-            const string inputCode =
-@"Property Get Foo$()
-End Property";
-
-            const string expectedCode =
-@"Sub Foo()
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new NonReturningFunctionInspection(state);
-            var inspectionResults = inspection.GetInspectionResults();
-
-            new ConvertToProcedureQuickFix(state).Fix(inspectionResults.First());
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void GivenImplicitVariantPropertyGetter_StillConvertsToSub()
-        {
-            const string inputCode =
-@"Property Get Foo()
-End Property";
-
-            const string expectedCode =
-@"Sub Foo()
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new NonReturningFunctionInspection(state);
-            var inspectionResults = inspection.GetInspectionResults();
-
-            new ConvertToProcedureQuickFix(state).Fix(inspectionResults.First());
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void GivenParameterizedPropertyGetter_QuickFixKeepsParameter()
-        {
-            const string inputCode =
-@"Property Get Foo(ByVal b As Boolean) As String
-End Property";
-
-            const string expectedCode =
-@"Sub Foo(ByVal b As Boolean)
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new NonReturningFunctionInspection(state);
-            var inspectionResults = inspection.GetInspectionResults();
-
-            new ConvertToProcedureQuickFix(state).Fix(inspectionResults.First());
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void NonReturningFunction_IgnoreQuickFixWorks()
-        {
-            const string inputCode =
-@"Function Foo() As Boolean
-End Function";
-
-            const string expectedCode =
-@"'@Ignore NonReturningFunction
-Function Foo() As Boolean
-End Function";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new NonReturningFunctionInspection(state);
-            var inspectionResults = inspection.GetInspectionResults();
-            
-            new IgnoreOnceQuickFix(state, new[] {inspection}).Fix(inspectionResults.First());
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
         }
 
         [TestMethod]
